refactor: LeverUpManger 개선
Enemy와의 직접 연동 해제
레벨 업데이트 최적화

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -15,7 +15,7 @@ public class Enemy : MonoBehaviour, IPoolable
     public EnemyAttackBase attackScript;
 
     public static event Action<Enemy> OnEnemyDeadGlobal;
-    public static event Action<float> OnEnemyExpGained;
+    public static event Action<int> OnEnemyExpGained;
 
 
     public float currentHP;
